requirements.txt
fastapi
uvicorn
sqlalchemy
aioodbc
pyodbc
python-dotenv
greenlet
starlette
pydantic
requests
selenium
apscheduler
PyPDF2
python-docx
python-multipart
openai
PyYAML
markdown-pdf
python-jose[cryptography]
httpx
email-validator
msal
pytest
pytest-asyncio
redis
rq
pytest-mock
aiosqlite
aiosqlite
